feat: adjust reports
