+++ src/components/containers/shop/reports/orderExpense/CardExpense.vue
@@ -112,7 +112,7 @@
   </div>
 </template>
 <script>
-import AppCardCapsule from '../../../modules/AppCardCapsule'
+import AppCardCapsule from '../../../../modules/AppCardCapsule'
 
 export default {
   name: 'AppCardExpense',
